requirements.txt
requests
openai
pytest
python-dotenv
chromadb